generate Rdb code into separate files
